# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillBridge AI",
    "description": "AI-powered career guidance for India's workforce",
    "version": "1.0.0"
  },
  "paths": {
...
    "/chat/message": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "operationId": "send_message_chat_message_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/history": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get History",
        "operationId": "get_history_chat_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_ChatMessageResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Clear History",
        "operationId": "clear_history_chat_history_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/assessment/start": {
      "post": {
        "tags": [
          "Assessment"
        ],
        "summary": "Start Assessment",
        "operationId": "start_assessment_assessment_start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_StartAssessmentResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assessment/answer": {
      "post": {
        "tags": [
          "Assessment"
        ],
        "summary": "Submit Answer",
        "operationId": "submit_answer_assessment_answer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AnswerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assessment/status": {
      "get": {
        "tags": [
          "Assessment"
        ],
        "summary": "Get Status",
        "operationId": "get_status_assessment_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AssessmentStatusResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assessment/history": {
      "get": {
        "tags": [
          "Assessment"
        ],
        "summary": "Get History",
        "operationId": "get_history_assessment_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_List_AssessmentHistoryItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_AnswerResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[AnswerResponse]"
      },
      "APIResponse_AssessmentStatusResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentStatusResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[AssessmentStatusResponse]"
      },
...
      "APIResponse_List_AssessmentHistoryItem__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssessmentHistoryItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[List[AssessmentHistoryItem]]"
      },
...
      "APIResponse_StartAssessmentResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StartAssessmentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[StartAssessmentResponse]"
      },
...
      "APIResponse_list_ChatMessageResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[list[ChatMessageResponse]]"
      },
...
      "AnswerResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "batch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentBatchResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "phase": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "question_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number"
          },
          "is_complete": {
            "type": "boolean",
            "title": "Is Complete"
          },
          "retakes_remaining": {
            "type": "integer",
            "title": "Retakes Remaining"
          },
          "skills_found": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills Found"
          },
          "career_goals": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Career Goals"
          },
          "assessment_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Summary"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "is_complete",
          "retakes_remaining"
        ],
        "title": "AnswerResponse",
        "description": "Feedback after submitting an answer."
      },
      "AssessmentAnswerRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Active session ID"
          },
          "answer": {
            "title": "Answer",
            "description": "User answer(s)"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "answer"
        ],
        "title": "AssessmentAnswerRequest",
        "description": "Request to submit assessment answers. Can be a single string or a list (for batches)."
      },
      "AssessmentBatchResponse": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__response__assessment__QuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          },
          "phase": {
            "type": "integer",
            "title": "Phase"
          },
          "phase_name": {
            "type": "string",
            "title": "Phase Name"
          }
        },
        "type": "object",
        "required": [
          "questions",
          "phase",
          "phase_name"
        ],
        "title": "AssessmentBatchResponse",
        "description": "A batch of questions returned by the adaptive engine."
      },
      "AssessmentHistoryItem": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "retake_number": {
            "type": "integer",
            "title": "Retake Number"
          },
          "is_complete": {
            "type": "boolean",
            "title": "Is Complete"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "skills_count": {
            "type": "integer",
            "title": "Skills Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "retake_number",
          "is_complete",
          "skills_count",
          "created_at"
        ],
        "title": "AssessmentHistoryItem",
        "description": "Summary of a past assessment session."
      },
      "AssessmentStatusResponse": {
        "properties": {
          "has_completed": {
            "type": "boolean",
            "title": "Has Completed"
          },
          "retakes_used": {
            "type": "integer",
            "title": "Retakes Used"
          },
          "retakes_remaining": {
            "type": "integer",
            "title": "Retakes Remaining"
          },
          "max_retakes": {
            "type": "integer",
            "title": "Max Retakes"
          },
          "last_completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Completed At"
          },
          "can_retake": {
            "type": "boolean",
            "title": "Can Retake"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "next_retake_available_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Retake Available At"
          },
          "has_incomplete": {
            "type": "boolean",
            "title": "Has Incomplete"
          },
          "incomplete_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incomplete Session Id"
          }
        },
        "type": "object",
        "required": [
          "has_completed",
          "retakes_used",
          "retakes_remaining",
          "max_retakes",
          "can_retake",
          "eligible",
          "has_incomplete"
        ],
        "title": "AssessmentStatusResponse",
        "description": "Current assessment status and eligibility for user."
      },
...
      "ChatMessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "language",
          "created_at"
        ],
        "title": "ChatMessageResponse",
        "description": "A single chat message in history.",
        "example": {
          "content": "You can start by learning SEO and SEM basics.",
          "created_at": "2024-05-01T10:00:00Z",
          "id": "msg_123",
          "language": "english",
          "role": "assistant"
        }
      },
      "ChatRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "language": {
            "$ref": "#/components/schemas/Language",
            "default": "english"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatRequest",
        "description": "Request message for AI chat.",
        "example": {
          "content": "How can I become a digital marketing expert?",
          "language": "english"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Language": {
        "type": "string",
        "enum": [
          "hindi",
          "english",
          "bengali",
          "telugu",
          "marathi",
          "tamil",
          "urdu",
          "gujarati",
          "kannada",
          "odia",
          "punjabi",
          "malayalam",
          "hinglish"
        ],
        "title": "Language"
      },
...
      "StartAssessmentResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "batch": {
            "$ref": "#/components/schemas/AssessmentBatchResponse"
          },
          "phase": {
            "type": "integer",
            "title": "Phase"
          },
          "phase_name": {
            "type": "string",
            "title": "Phase Name"
          },
          "question_number": {
            "type": "integer",
            "title": "Question Number"
          },
          "retakes_used": {
            "type": "integer",
            "title": "Retakes Used"
          },
          "retakes_remaining": {
            "type": "integer",
            "title": "Retakes Remaining"
          },
          "max_retakes": {
            "type": "integer",
            "title": "Max Retakes"
          },
          "can_resume": {
            "type": "boolean",
            "title": "Can Resume"
          },
          "is_complete": {
            "type": "boolean",
            "title": "Is Complete",
            "default": false
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible",
            "default": true
          },
          "has_incomplete": {
            "type": "boolean",
            "title": "Has Incomplete",
            "default": false
          },
          "incomplete_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incomplete Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "batch",
          "phase",
          "phase_name",
          "question_number",
          "retakes_used",
          "retakes_remaining",
          "max_retakes",
          "can_resume"
        ],
        "title": "StartAssessmentResponse",
        "description": "Response after starting/resuming an assessment."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__response__assessment__QuestionResponse": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type",
            "default": "mcq"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "allows_multiple": {
            "type": "boolean",
            "title": "Allows Multiple",
            "default": false
          },
          "allows_other": {
            "type": "boolean",
            "title": "Allows Other",
            "default": true
          },
          "skill_probing": {
            "type": "string",
            "title": "Skill Probing"
          }
        },
        "type": "object",
        "required": [
          "question",
          "skill_probing"
        ],
        "title": "QuestionResponse",
        "description": "Response containing a single question within a batch."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}